Simplification pass + frame-builder size fix
